+++ com.unity.render-pipelines.universal/Runtime/2D/CompositeLightReactor2D.cs
@@ -1,0 +1,22 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+
+namespace UnityEngine.Experimental.Rendering.Universal
+{
+
+    [ExecuteInEditMode]
+    public class CompositeLightReactor2D : ShadowCasterGroup2D
+    {
+        protected void OnEnable()
+        {
+            ShadowCasterGroup2DManager.AddGroup(this);
+        }
+
+        protected void OnDisable()
+        {
+            ShadowCasterGroup2DManager.RemoveGroup(this);
+        }
+    }
+}