+++ com.unity.render-pipelines.universal/Runtime/2D/Passes/Render2DLightingPass.cs
@@ -14,8 +14,7 @@ namespace UnityEngine.Experimental.Rendering.Universal
         static readonly ShaderTagId k_NormalsRenderingPassName = new ShaderTagId("NormalsRendering");
         static readonly ShaderTagId k_LegacyPassName = new ShaderTagId("SRPDefaultUnlit");
         static readonly List<ShaderTagId> k_ShaderTags = new List<ShaderTagId>() { k_LegacyPassName, k_CombinedRenderingPassName, k_CombinedRenderingPassNameOld };
-        //static readonly List<ShaderTagId> k_ShaderTags = new List<ShaderTagId>() { k_CombinedRenderingPassName };
-
+        
         public Render2DLightingPass(Renderer2DData rendererData)
         {
             if (s_SortingLayers == null)
@@ -104,7 +103,7 @@ namespace UnityEngine.Experimental.Rendering.Universal
 #if UNITY_EDITOR
                     cmd.name = "Render Light Volumes" + SortingLayer.IDToName(layerToRender);
 #endif
-                    RendererLighting.RenderLightVolumes(camera, cmd, layerToRender);
+                    RendererLighting.RenderLightVolumes(camera, cmd, layerToRender, colorAttachment);
                     context.ExecuteCommandBuffer(cmd);
                     cmd.Clear();
                 }