+++ com.unity.render-pipelines.universal/Runtime/2D/Renderer2DData.cs
@@ -36,18 +36,32 @@ namespace UnityEngine.Experimental.Rendering.Universal
         [SerializeField]
         Shader m_BlitShader = null;
 
+        [SerializeField]
+        Shader m_ShadowShader = null;
+
+        [SerializeField]
+        Shader m_ShadowGroupShader = null;
+
+        [SerializeField]
+        Shader m_RemoveSelfShadowShader = null;
+
         [SerializeField, Reload("Runtime/Data/PostProcessData.asset")]
         PostProcessData m_PostProcessData = null;
 
         public float hdrEmulationScale => m_HDREmulationScale;
         public Light2DBlendStyle[] lightBlendStyles => m_LightBlendStyles;
 
+
         internal Shader shapeLightShader => m_ShapeLightShader;
         internal Shader shapeLightVolumeShader => m_ShapeLightVolumeShader;
         internal Shader pointLightShader => m_PointLightShader;
         internal Shader pointLightVolumeShader => m_PointLightVolumeShader;
         internal Shader blitShader => m_BlitShader;
+        internal Shader shadowShader => m_ShadowShader;
+        internal Shader shadowGroupShader => m_ShadowGroupShader;
+        internal Shader removeSelfShadowShader => m_RemoveSelfShadowShader;
         internal PostProcessData postProcessData => m_PostProcessData;
+
 
         protected override ScriptableRenderer Create()
         {
@@ -106,6 +120,9 @@ namespace UnityEngine.Experimental.Rendering.Universal
             m_PointLightShader = Shader.Find("Hidden/Light2D-Point");
             m_PointLightVolumeShader = Shader.Find("Hidden/Light2d-Point-Volumetric");
             m_BlitShader = Shader.Find("Hidden/Universal Render Pipeline/Blit");
+            m_ShadowShader = Shader.Find("Hidden/Shadow2D");
+            m_ShadowGroupShader = Shader.Find("Hidden/ShadowGroup2D");
+            m_RemoveSelfShadowShader = Shader.Find("Hidden/Shadow2DRemoveSelf");
         }
 
         protected override void OnEnable()